--- test/test.rb
@@ -1,4 +1,0 @@
-require 'minitest'
-require 'minitest/autorun'
-require 'minitest/pride'
-require_relative '../lib/melon'
